% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose_msts}
\alias{decompose_msts}
\title{decompose msts}
\usage{
decompose_msts(df, .value, .seasonal_periods = 24)
}
\arguments{
\item{df}{a dataframe}

\item{.value}{a target variable}

\item{.seasonal_periods}{a single or vector of seasonal periods}
}
\description{
decompose msts
}
